Fixed #4 - Used proper datatypes for values

+++ src/Sudoku.h
@@ -9,6 +9,7 @@
 #include <utility>
 #include <ctime>
 #include <cstdlib>
+#include <cstdint>
 #include <cstring>
 
 class Sudoku
@@ -24,17 +25,17 @@ class Sudoku
     Sudoku() = default;
 
     /* Parameterized constructor that takes the initial state of the board as input*/
-    explicit Sudoku(const int (&board)[9][9]);
+    explicit Sudoku(const uint8_t (&board)[9][9]);
 
     /* A function that checks if the current board configuration is valid
-         * i.e., it checks if there are any repetitions in the rows/cols/box */
+     * i.e., it checks if there are any repetitions in the rows/cols/box */
     [[nodiscard]] bool is_valid_configuration() const noexcept;
 
     /* A function that returns the location of the first empty position */
-    [[nodiscard]] std::pair<int, int> get_first_empty() const noexcept;
+    [[nodiscard]] std::pair<size_t, size_t> get_first_empty() const noexcept;
 
     /* A function that checks if all the cells of the board are non-empty
-         * and if the current configuration of the board is valid */
+     * and if the current configuration of the board is valid */
     [[nodiscard]] bool is_solved() const noexcept;
 
     /* A function that generates a random sudoku board with a fixed number of clues */
@@ -47,7 +48,7 @@ class Sudoku
     void print_sudoku() const noexcept;
 
   private:
-    int board[9][9] = { 0 };
+    uint8_t board[9][9] = { 0 };
 };
 
 #endif
